requirements.txt
# For Apple Silicon Macs
tensorflow-macos>=2.13.0
tensorflow-metal>=1.0.0
flask>=2.0.0
pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0